Refactor function locations; empty datasets and sessions frames with correct dtypes
